% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puny.R
\name{code}
\alias{code}
\title{punycode encoder}
\usage{
code(s)
}
\arguments{
\item{s}{a character vector}
}
\value{
s encoded in \href{https://en.wikipedia.org/wiki/Punycode}{punycode}
}
\description{
punycode encoder
}
\examples{

# cedilla
puny::code( "François" )

# emojis
s <- emo::ji_glue("hello:poop:")
puny::code(s)

# more french
puny::code( "crème brûlée" )

}
