% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puny.R
\name{decode}
\alias{decode}
\title{punycode decoder}
\usage{
decode(s)
}
\arguments{
\item{s}{a character vector}
}
\value{
a utf-8 string decoded from \code{s}
}
\description{
Decodes a string encoded in \href{https://en.wikipedia.org/wiki/Punycode}{punycode}
}
